Clean up time in bed metric
